+++ boards/arm64/qemu/qemu-armv8a/configs/nsh/defconfig
@@ -7,8 +7,8 @@
 #
 CONFIG_ARCH="arm64"
 CONFIG_ARCH_ARM64=y
-CONFIG_ARCH_BOARD="qemu-a53"
-CONFIG_ARCH_BOARD_QEMU_A53=y
+CONFIG_ARCH_BOARD="qemu-armv8a"
+CONFIG_ARCH_BOARD_QEMU_ARMV8A=y
 CONFIG_ARCH_CHIP="qemu"
 CONFIG_ARCH_CHIP_QEMU=y
 CONFIG_ARCH_CHIP_QEMU_A53=y
