+++ boards/arm64/qemu/qemu-armv8a/scripts/Make.defs
@@ -1,5 +1,5 @@
 ############################################################################
-# boards/arm64/qemu/qemu-a53/scripts/Make.defs
+# boards/arm64/qemu/qemu-armv8a/scripts/Make.defs
 #
 # Licensed to the Apache Software Foundation (ASF) under one or more
 # contributor license agreements.  See the NOTICE file distributed with
