+++ boards/arm64/qemu/qemu-armv8a/src/Makefile
@@ -1,5 +1,5 @@
 ############################################################################
-# boards/arm64/qemu/qemu-a53/src/Makefile
+# boards/arm64/qemu/qemu-armv8a/src/Makefile
 #
 # Licensed to the Apache Software Foundation (ASF) under one or more
 # contributor license agreements.  See the NOTICE file distributed with
