+++ boards/arm64/qemu/qemu-armv8a/src/qemu_appinit.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * boards/arm64/qemu/qemu-a53/src/qemu_bringup.c
+ * boards/arm64/qemu/qemu-armv8a/src/qemu_appinit.c
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
@@ -25,38 +25,52 @@
 #include <nuttx/config.h>
 
 #include <sys/types.h>
-#include <syslog.h>
 
-#include <nuttx/fs/fs.h>
+#include <nuttx/board.h>
 
-#include "qemu-a53.h"
+#include "qemu-armv8a.h"
+
+#ifdef CONFIG_BOARDCTL
 
 /****************************************************************************
  * Public Functions
  ****************************************************************************/
 
 /****************************************************************************
- * Name: imx_bringup
+ * Name: board_app_initialize
  *
  * Description:
- *   Bring up board features
+ *   Perform application specific initialization.  This function is never
+ *   called directly from application code, but only indirectly via the
+ *   (non-standard) boardctl() interface using the command BOARDIOC_INIT.
+ *
+ * Input Parameters:
+ *   arg - The boardctl() argument is passed to the board_app_initialize()
+ *         implementation without modification.  The argument has no
+ *         meaning to NuttX; the meaning of the argument is a contract
+ *         between the board-specific initialization logic and the
+ *         matching application logic.  The value could be such things as a
+ *         mode enumeration value, a set of DIP switch switch settings, a
+ *         pointer to configuration data read from a file or serial FLASH,
+ *         or whatever you would like to do with it.  Every implementation
+ *         should accept zero/NULL as a default configuration.
+ *
+ * Returned Value:
+ *   Zero (OK) is returned on success; a negated errno value is returned on
+ *   any failure to indicate the nature of the failure.
  *
  ****************************************************************************/
 
-int qemu_bringup(void)
+int board_app_initialize(uintptr_t arg)
 {
-  int ret;
+  UNUSED(arg);
+#ifndef CONFIG_BOARD_LATE_INITIALIZE
+  /* Perform board initialization */
 
-#ifdef CONFIG_FS_PROCFS
-  /* Mount the procfs file system */
-
-  ret = nx_mount(NULL, "/proc", "procfs", 0, NULL);
-  if (ret < 0)
-    {
-      syslog(LOG_ERR, "ERROR: Failed to mount procfs at /proc: %d\n", ret);
-    }
-#endif
-
-  UNUSED(ret);
+  return qemu_bringup();
+#else
   return OK;
+#endif
 }
+
+#endif /* CONFIG_BOARDCTL */
