+++ boards/arm64/qemu/qemu-armv8a/src/qemu_bringup.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * boards/arm64/qemu/qemu-a53/src/qemu_appinit.c
+ * boards/arm64/qemu/qemu-armv8a/src/qemu_bringup.c
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
@@ -25,52 +25,38 @@
 #include <nuttx/config.h>
 
 #include <sys/types.h>
+#include <syslog.h>
 
-#include <nuttx/board.h>
+#include <nuttx/fs/fs.h>
 
-#include "qemu-a53.h"
-
-#ifdef CONFIG_BOARDCTL
+#include "qemu-armv8a.h"
 
 /****************************************************************************
  * Public Functions
  ****************************************************************************/
 
 /****************************************************************************
- * Name: board_app_initialize
+ * Name: imx_bringup
  *
  * Description:
- *   Perform application specific initialization.  This function is never
- *   called directly from application code, but only indirectly via the
- *   (non-standard) boardctl() interface using the command BOARDIOC_INIT.
- *
- * Input Parameters:
- *   arg - The boardctl() argument is passed to the board_app_initialize()
- *         implementation without modification.  The argument has no
- *         meaning to NuttX; the meaning of the argument is a contract
- *         between the board-specific initialization logic and the
- *         matching application logic.  The value could be such things as a
- *         mode enumeration value, a set of DIP switch switch settings, a
- *         pointer to configuration data read from a file or serial FLASH,
- *         or whatever you would like to do with it.  Every implementation
- *         should accept zero/NULL as a default configuration.
- *
- * Returned Value:
- *   Zero (OK) is returned on success; a negated errno value is returned on
- *   any failure to indicate the nature of the failure.
+ *   Bring up board features
  *
  ****************************************************************************/
 
-int board_app_initialize(uintptr_t arg)
+int qemu_bringup(void)
 {
-  UNUSED(arg);
-#ifndef CONFIG_BOARD_LATE_INITIALIZE
-  /* Perform board initialization */
+  int ret;
 
-  return qemu_bringup();
-#else
-  return OK;
+#ifdef CONFIG_FS_PROCFS
+  /* Mount the procfs file system */
+
+  ret = nx_mount(NULL, "/proc", "procfs", 0, NULL);
+  if (ret < 0)
+    {
+      syslog(LOG_ERR, "ERROR: Failed to mount procfs at /proc: %d\n", ret);
+    }
 #endif
-}
 
-#endif /* CONFIG_BOARDCTL */
+  UNUSED(ret);
+  return OK;
+}
